requirements.txt
2to3==1.0
asttokens==2.1.0
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
brotlipy==0.7.0
charset-normalizer==2.1.1
click==8.1.3
cog==0.0.3
cytoolz==0.11.0
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
filelock==3.8.0
fonttools==4.25.0
ftfy==6.1.1
h11==0.14.0
huggingface-hub==0.9.1
idna==3.4
imgaug==0.4.0
ipython==8.6.0
jedi==0.18.1
jieba==0.42.1
joblib==1.2.0
matplotlib-inline==0.1.6
mkl-service==2.4.0
munkres==1.1.4
numpy==1.22.0
opencv-python==4.8.0.74
outcome==1.2.0
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptflops==0.7
pure-eval==0.2.2
Pygments==2.13.0
pynvml==11.4.1
pytz==2023.3
PyYAML==6.0
regex==2022.9.13
sacremoses==0.0.53
scikit-image==0.18.1
selenium==4.7.2
Shapely==1.8.5.post1
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.0
thop==0.1.1.post2209072238
torchaudio==0.12.1
torchinfo==1.8.0
torchstat==0.0.7
torchsummary==1.5.1
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2023.3
urllib3==1.26.12
wcwidth==0.2.5
wincertstore==0.2
wsproto==1.2.0
